Returns cooldown reduction
Based on conversations with various community members, we've come to a compromise on cooldown reduction. With a 3 month lock up and an expected earliest implementation of a chain migration at the end of the year, this should allow mining pools to make a short term decision on unstaking and returning funds to constituents.

We also remove the need for vesting HNT as all staked HNT now becomes veHNT after the chain migration.