Initial Draft of HIP 71
This is based on HIP 70, but proposes moving to Kujira, an L1 Cosmos SDK based chain that supports IBC and has scaled rapidly in the last several months.  This ultimately presents a greater value proposition for HNT, IOT, and MOBILE miners and allows them to easily participate in the most exciting development space in crypto today.