Addressing Concerns
We've begun to clarify the areas that have been noted. An additional major edit is underway; respecifying DC usage in place of BOP, the dissolution of the BOP token into a dVPN subDAO treasury, a $VPN coin, veVPN, and the mechanisms that will be used to implement these solutions. Further detail into the pricing mechanisms behind dVPN payment and rewards structure using a $VPN/DC model, as well as providing relevant code as it becomes available. These edits will take more time as we develop the mechanisms proposed. Thanks to all for the feedback on this draft.