Introduce new DAO Utility Score (#597)
* Introduce new DAO Utility Score

After discussions with the HIP-51 team, a new simplified calculation of the DAO Utility Score is presented that better aligns the interests of DAO participants.

* Introduce new DAO Utility Score

After discussions with the HIP-51 team, a new simplified calculation of the DAO Utility Score is presented that better aligns the interests of DAO participants. (+ minor updates)

* Introduce new DAO Utility Score

After discussions with the HIP-51 team, a new simplified calculation of the DAO Utility Score is presented that better aligns the interests of DAO participants. (+ minor updates)

* add hip 78 back and create hip 80

* change mentions of hip 78 to 80

* Update 0080-updating-dao-utility-score.md

* rename file

* copying changes from other branch

---------

Co-authored-by: vincenzospaghetti <93289575+vincenzospaghetti@users.noreply.github.com>