Allow a wallet user to onboard with own funds if maker leaves network
This HIP proposes a mechanism to improve the onboarding experience for hotspot users whose makers have left the network either voluntarily or through bankruptcy. Currently, if a maker leaves the network and has run out of onboarding credits, the users are left with only cumbersome or risky ways to onboard their devices to the network. This proposal seeks to change this in a way that lets users onboard their devices in a self-service manner if their hotspot maker has left the ecosystem.