Update 0087-proportional-service-provider-rewards.md (#756)
Remove the unresolved issue of whether HIP-79 passed or not.  It had no impact but is no longer necessary to even mention.  Also corrected the format when a HIP is mentioned (HIP-XX).  Maybe just a few grammatical improvements.  No impact has been made to the content, ideas, or points covered; as such, there is no need to reset the 7-day waiting period.