Create Setting-CPI-Requirements-CBRS-Installations.md
This HIP aims to set CPI (Certified Professional Installer) requirements for CBRS deployments to better the installation quality and knowledge of the deployers for the CBRS network. 
Every new cbrs deployment and/or re-registration of a cbrs radios, and will require a CPI identification number when submitting for SAS registration.