0122-amend-service-provider-hex-boosting.md
This HIP aims to do three things to amend HIP-84 Service Provider Hex Boosting. It creates a dedicated Service Provider Hex Boosting emissions funding category. It reduces the minimum time for Service Provider Hex Boosting to 3 months and finally limits future boost creation to a maximum of 10x rewards.

The dedicated Service Provider Hex Boosting Rewards bucket will be limited to a maximum of 10% of all future emissions in the MOBILE network and will be taken out of the current Proof-of-Coverage category (reduced from 20% to a minimum of 10%). Any of the 10% not used by Service Provider Hex Boosting is returned to the Proof-of-Coverage bucket.
The minimum time for Service Provider Hex Boosting is reduced to 3 months from 6 months to allow more flexibility of boosting to hexes that may have an unsure value to the network.
Limit FUTURE hex boost creation to a maximum of 10x rewards. As the community believes the current 100x boosted hexes are breeding temporary deployments that are reward-chasing and attempting gaming of Proof-of-Coverage.