Create 0000-Flow-through-treasury-for-conducting-daily-HNT-sales-to-actively-burn-MOBILE.md
The proposal is to use the emissions of HNT into the MOBILE treasury for automated daily auctions. The daily automated auctions sell HNT for MOBILE. All MOBILE purchased by the treasury in this way are subsequently burned. The proposed volume of the daily auctions is 100% of the HNT emissions from the most recent epoch and a portion (e.g., 0.25%) of the total remaining treasury account.